% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network3d.R
\name{network3d}
\alias{network3d}
\title{network3d}
\usage{
network3d(vertices, edges, camera = list(), controls = list(),
  node_outline_black = TRUE, background_color = "white", node_size = 0.1,
  raycast_res = 0.05, edge_color = 12237498, edge_opacity = 0.1,
  interactive = TRUE, selection_size_mult = 1.5, tooltip_offset = 15,
  select_all = TRUE, show_simulation_progress = TRUE,
  max_iterations = 100, manybody_strength = -1, link_strength = NULL,
  static_length_strength = FALSE, force_explorer = FALSE, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{vertices}{Dataframe with at least one column: id, optionally a 'color' column with css valid colors of nodes, 'size' column with sizes of each vertice, 'name' column for text of mouseover tooltip, and 'interactive' boolean column for if the node can be interacted with or not.}

\item{edges}{Dataframe with two columns: 'source' or the id of the node edge is coming from, and 'target' or id of node edge is going to.}

\item{node_outline_black}{Outline the node circles in black? Default (FALSE) is white.}

\item{background_color}{Color of background of plot. Any css valid color will work.}

\item{node_size}{How big should the nodes be? Relative to world size of 2x2x2. Note that this is overwritten if the vertices dataframe has a size column.}

\item{raycast_res}{Thickness of invisible raycasting selection beam. Bigger values will make it easier to select but will cause more misselections.}

\item{edge_color}{Color of lines conencting node/vertices.}

\item{edge_opacity}{Transparency of lines connecting node/vertices.}

\item{interactive}{Is the network interactive? I.e. does mousing over a node display what's in its 'name' column?}

\item{selection_size_mult}{How much do moused over nodes get expanded?}

\item{tooltip_offset}{Tooltip that shows whatever's in the 'name' field should be offset by how much? (this is in screen pixels) Too little and the tip obscures your nodes, too much and it can be hard to tell what you've selected.}

\item{select_all}{Do we show tooltip for every node or just nodes with TRUE in a column 'selectable' in the node values?}

\item{show_simulation_progress}{Show small popup while layout is being calculated so user knows when it has finished?}

\item{max_iterations}{Number of iterations the layout simulation runs.}

\item{manybody_strength}{Attractive force between nodes irrespective of links. See https://github.com/d3/d3-force#many-body for more details/}

\item{link_strength}{attractive force of links. Falsy values default to a function of number of connections. See https://github.com/d3/d3-force#links for more details.}

\item{static_length_strength}{If TRUE then all links have same strength, otherwise they are scaled by the number of connections each node has. (More connections = each link is less strong).}

\item{manybody_strength}{How strong is the pull between nodes regardless of links? Positive means they are attracted to eachother, negative: repelled.}

\item{link_strength}{How strong are the links between the nodes?}
}
\description{
Render a 3d network visualization in an htmlwidget. Calculates the layout simulation within javascript and is fast and lightweight.
}
\examples{

# Basic use with a custom number of simulation iterations
network3d::network3d(data, max_iterations = 75)

# Spins up explorer to fiddle with simulation parameters
network3d::network3d(data, max_iterations = 75, manybody_strength = 1, force_explorer = TRUE)

# Custom force parameters
network3d::network3d(data, max_iterations = 75, manybody_strength = 1)

# Custom camera settings: sets camera way far away from graph.
network3d::network3d(
  data,
  camera = list(
    start_pos = list(x=1.2, y=1.2, z=20)
  ),
  max_iterations = 75
)

}
